subscriber/requirements.txt
psycopg2-binary
pytz
paho-mqtt